Move QT designer files into subdir and ui files into module. Remove unused imports and unused code.
