+++ designer/changeCallsignDialog.ui
