+++ designer/clueDialog.ui
