+++ designer/clueLogDialog.ui
