+++ designer/fsFilterDialog.ui
