+++ designer/fsSendMessageDialog.ui
