+++ designer/newEntryWidget.ui
