+++ designer/newEntryWindow.ui
