+++ designer/nonRadioClueDialog.ui
