+++ designer/opPeriodDialog.ui
