+++ designer/printClueLogDialog.ui
