+++ designer/subjectLocatedDialog.ui
