+++ ui/changeCallsignDialog_ui.py
