+++ ui/clueLogDialog_ui.py
