+++ ui/newEntryWidget_ui.py
