+++ ui/newEntryWindow_ui.py
