+++ ui/printClueLogDialog_ui.py
